Move Predicates util to com.facebook.presto.util
